Added Recommended Courses. First major bug
Added the ability to add course recommendations to a Career. Experiencing an error passing the PDO object to a function, as it is expecting a string(?) for some reason and PHP is reading it as an array. While making my PDO Global is a quick and dirty fix, i am working on finding a cleaner solution.

added footer nav bar for authenticated users